requirements.txt
Django==4.2.9
djangorestframework==3.14.0
celery[redis]==5.3.4
redis==5.0.1
mysqlclient==2.2.0
openai==1.6.1
sentence-transformers==2.2.2
torch==2.1.0
requests==2.31.0
Pillow==10.1.0
langdetect==1.0.9
django-crontab==0.7.1
django-extensions==3.2.3
numpy==1.24.3
scipy==1.11.4
python-dotenv==1.0.1

# Developer tools
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==5.0.0
ruff==0.6.9
mypy==1.12.0
